rental vehicle done

+++ RentalService/src/main/java/com/rent/RentalService/repository/RentVehicleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.rent.RentalService.model.RentVehicle;
 
 @Repository
-public interface RentVehicleRepository extends JpaRepository<RentVehicle, Integer> {
+public interface RentVehicleRepository extends JpaRepository<RentVehicle, Long> {
 
 }
